+++ CHANGELOG.md
@@ -25,6 +25,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - [Map form element #required is not enforced #560](https://github.com/farmOS/farmOS/issues/560)
 - [Incorrect translation placeholders for asset names #540](https://github.com/farmOS/farmOS/issues/540)
 - Update farmOS-map to [v2.0.5](https://github.com/farmOS/farmOS-map/releases/tag/v2.0.5) to fix [Uncaught (in promise) TypeError: o.getChangeEventType is not a function #551](https://github.com/farmOS/farmOS/issues/551)
+- [Prevent circular group membership #562](https://github.com/farmOS/farmOS/pull/562)
 
 ## [2.0.0-beta6] 2022-07-30